+++ docs/source/release-notes.rst
@@ -15,6 +15,8 @@ v1.9.4 (2025-xx-xx)
   list.
 - Fix a bug in line model that causes incorrect admittance matrix.
 - Correct a typo in the PSSE DYR parser, changing "Tprod" to "Tpord".
+- Fix a bug in PSSE RAW parser in incorrect handling of 3WT. Also fixes issue in
+  handling quoted strings with commas and slashes.
 
 v1.9.3 (2025-01-05)
 -------------------